update state issues
